+++ src/com/skilldistillery/entities/ResearchJet.java
@@ -2,7 +2,7 @@ package com.skilldistillery.entities;
 
 public class ResearchJet extends Jet implements IResearch {
 	
-	public ResearchJet(String model, int speed, int range, long price) {
+	public ResearchJet(String model, double speed, int range, long price) {
 		this.model = model;
 		this.speed=speed;
 		this.range = range;
